Fix fsync and metadata search
fsync: handle block buffer contents not fitting in data block
find_data_block: handle cases with fsynced writes < data_block_size,
where our metadata entries are further ahead than the computed expected
locations

+++ tests/basic.c
@@ -120,7 +120,7 @@ int main(int argc, char **argv) {
     {
         const char *filename = "/testfile";
         int ret = r2f2_write(&fs, fd, write_buf, write_len);
-        if (ret != write_len) {
+        if (ret != (int)write_len) {
             printf("R2F2 write '%s' to file '%s' failed (%d)\n", write_buf,
                    filename, ret);
             return ret;
@@ -146,7 +146,7 @@ int main(int argc, char **argv) {
         uint8_t read_buf[write_len];
         memset(read_buf, 0, sizeof(read_buf));
         int ret = r2f2_read(&fs, fd, read_buf, write_len);
-        if (ret != write_len) {
+        if (ret != (int)write_len) {
             printf("R2F2 read %zu B from file '%s' failed (%d)\n",
                    strlen(write_buf) + 1, filename, ret);
             return ret;
